+++ SpecFlow.VisualStudio.Package/ProjectSystem/NullVsIdeScope.cs
@@ -219,7 +219,7 @@ namespace SpecFlow.VisualStudio.ProjectSystem
             return null;
         }
 
-        public Task RunOnBackgroundThread(Func<Task> action, Action<Exception> onException)
+        public Task RunOnBackgroundThread(Func<Task> action, Action<Exception> onException, string threadName)
         {
             return Task.CompletedTask;
         }