+++ SpecFlow.VisualStudio/ProjectSystem/IIdeScope.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO.Abstractions;
+using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using SpecFlow.VisualStudio.Diagnostics;
@@ -30,7 +31,7 @@ namespace SpecFlow.VisualStudio.ProjectSystem
         IDisposable CreateUndoContext(string undoLabel);
         bool GetTextBuffer(SourceLocation sourceLocation, out ITextBuffer textBuffer);
         SyntaxTree GetSyntaxTree(ITextBuffer textBuffer);
-        Task RunOnBackgroundThread(Func<Task> action, Action<Exception> onException);
+        Task RunOnBackgroundThread(Func<Task> action, Action<Exception> onException, string threadName);
         Task RunOnUiThread(Action action);
         void OpenIfNotOpened(string path);
     }