+++ Tests/SpecFlow.VisualStudio.Tests/ApprovalTestData/ReprocessStepDefinitionFileTests.Approval.IntParameter.cs
@@ -1,0 +1,15 @@
+using System;
+using TechTalk.SpecFlow;
+
+namespace TestProject
+{
+    [Binding]
+    public class Feature1StepDefinitions
+    {
+        [When(@"I press (.*)")]
+        public void WhenIPress(int intParameter)
+        {
+            throw new PendingStepException();
+        }
+    }
+}