+++ Tests/SpecFlow.VisualStudio.Tests/ApprovalTestData/ReprocessStepDefinitionFileTests.Approval.MultipleStepDefinitions.cs
@@ -1,0 +1,21 @@
+using System;
+using TechTalk.SpecFlow;
+
+namespace TestProject
+{
+    [Binding]
+    public class Feature1StepDefinitions
+    {
+        [When(@"I press add")]
+        public void WhenIPressAdd()
+        {
+            throw new PendingStepException();
+        }
+
+        [When(@"I press subtract")]
+        public void WhenIPressAdd()
+        {
+            throw new PendingStepException();
+        }
+    }
+}