+++ Tests/SpecFlow.VisualStudio.Tests/Editor/Services/DeveroomTaggerTests.cs
@@ -47,7 +47,7 @@ public class DeveroomTaggerTests
             _recalculateAction();
         }
 
-        public IEnumerable<LogMessage> LoggerMessages => ((IdeScope.Object.Logger as StubLogger)!).Messages;
+        public IEnumerable<LogMessage> LoggerMessages => ((IdeScope.Object.Logger as StubLogger)!).Logs;
         public IEnumerable<LogMessage> LoggerErrorMessages => LoggerMessages.Where(m=>m.Level == TraceLevel.Error || m.Message.Contains("Exception"));
 
         public void AssertNoErrorLogged() => LoggerErrorMessages.Should().BeEmpty();